pinocchio_log

Module logger

Macros

impl_log_for_signed 🔒

Implement the log trait for the signed integer types.

impl_log_for_slice 🔒

Implement the log trait for the slice type.

impl_log_for_unsigned_integer 🔒

Implement the log trait for unsigned integer types.

Structs

Logger

Logger to efficiently format log messages.

Enums

Argument

Formatting arguments.

Constants

TRUNCATED 🔒

Byte representing a truncated log.

TRUNCATED_SLICE 🔒

Bytes for a truncated str log message.

UNINIT_BYTE 🔒

An uninitialized byte.

Traits

Log

Trait to specify the log behavior for a type.

Functions

log_message

Log a message.